refactor: change bot commands manager export parameters
